+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 * *BREAKING* Align the `finance.*` interfaces versions in accordance with the changes in the descriptor. Refs UIF-520.
 * Assign locations to restricted funds in central tenant with central ordering enabled. Refs UIF-511.
 * Change rollover restriction lables. Refs UIF-537.
+* Display "Credited" total in Fiscal year view. Refs UIF-494.
+* Display "Credited" total in Ledger view .Refs UIF-495.
+* Display "Credited" total in Group view. Rfs UIF-496.
 
 ## [6.0.0](https://github.com/folio-org/ui-finance/tree/v6.0.0) (2024-03-19)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v5.0.0...v6.0.0)