+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Add filter to transactions view for "Encumbrance status". Refs UIF-535.
 * Add error handling for the `budgetRestrictedExpendituresError` error code. Refs UIF-543.
 * Display currency on Fiscal Year and Budget. Refs UIF-538.
+* Clean up permission names. Refs UIF-550.
 
 ## [6.0.0](https://github.com/folio-org/ui-finance/tree/v6.0.0) (2024-03-19)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v5.0.0...v6.0.0)